refactor: reorganize and update type imports for consistency; add raw sound and music data
