+++ src/components/controls/main/header.tsx
@@ -19,7 +19,7 @@ import { Slider } from '@/components/ui/slider'
 import { useAudio } from '@/context/audio'
 import { useSettings } from '@/context/setting'
 import { useSounds } from '@/context/sounds'
-import type { Settings } from '@/types'
+import type { Settings } from '@/types/settings'
 import { addTrack } from '@/utils/audio/addTrack'
 
 export function Header() {
