+++ src/context/setting.tsx
@@ -2,7 +2,7 @@ import type React from 'react'
 import { createContext, useCallback, useContext, useEffect, useState } from 'react'
 import { defaultSettings } from '@/constants/settings'
 
-import type { Settings } from '@/types'
+import type { Settings } from '@/types/settings'
 
 interface SettingsContextType {
   settings: Settings
