+++ src/types/theme.ts
@@ -1,0 +1,8 @@
+export interface ThemePreset {
+  name: string
+  backgroundColor: string
+  buttonColor: string
+  buttonText: string
+  buttonActive: string
+  buttonHoverColor: string
+}
